requirements.txt
azure-functions>=1.17.0
requests>=2.31.0
python-dotenv>=1.0.0
